feat: add @richiejp / VoxInput
The realtime support is not working due to an "unknown server sent event: transcription_session.created".
I can't get the GUI to show, either, however this is less of a concern.

I also need to add system-wide keybindings to start `voxinput`, record
immediately and then stop the daemon afterwards - I don't need it
hanging around indefinitely in the background.